refactor: Trasladar lógica de add/removeIngredient a ProductService
Los métodos de dominio (add/removeIngredient) en la entidad Product tomaban Foods como parámetro (alto acoplamiento). También lanzaban excepciones, lo que hacía que Product tuviese demasiadas responsabilidades.
Se ha pasado la creación de los Ingredients a ProductService. Los métodos de dominio en Product se han eliminado, porque el constructor y métodos de Ingredient ya se encargan de la consistencia bidireccional.

+++ app/spring-boot/src/main/java/es/ubu/inf/tfg/product/ProductController.java
@@ -13,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -111,7 +112,7 @@ public class ProductController {
     public ResponseEntity<IngredientResponseDTO> addIngredientToProduct(
             @PathVariable Integer productId,
             @RequestParam Integer foodId,
-            @RequestParam Double quantity) {
+            @RequestParam BigDecimal quantity) {
         IngredientResponseDTO ingredient = productService.addIngredientToProduct(productId, foodId, quantity);
         return ResponseEntity.status(HttpStatus.CREATED).body(ingredient);
     }
@@ -121,7 +122,7 @@ public class ProductController {
     public ResponseEntity<IngredientResponseDTO> updateIngredientQuantity(
             @PathVariable Integer productId,
             @PathVariable Integer foodId,
-            @RequestParam Double quantity) {
+            @RequestParam BigDecimal quantity) {
         IngredientResponseDTO ingredient = productService.updateIngredientQuantity(productId, foodId, quantity);
         return ResponseEntity.ok(ingredient);
     }
